add UserProfile model and admin inline Userprofile in User and add tests
